Microsecond sleep

+++ Kernel/Arch/aarch64/Include/Arch/timer.h
@@ -3,7 +3,7 @@
 
 namespace Timer::Arch
 {
-    size_t Ticks()
+    uint64 Ticks()
     {
         return 0;
     }
